requirements.txt
configobj
slackclient
nest_asyncio
croniter